Move job pages to job folder
