+++ apps/frontend/src/app/router.tsx
@@ -17,9 +17,9 @@ const LoginPage = lazy(() => import('../pages/Auth/Login'));
 const RegisterPage = lazy(() => import('../pages/Auth/Register'));
 
 // Import new pages and their loader/action functions
-const JobListPage = lazy(() => import('../pages/JobList'));
-const JobDetailPage = lazy(() => import('../pages/JobDetail'));
-import { jobDetailLoader, applyAction } from '../pages/JobDetail';
+const JobListPage = lazy(() => import('../pages/Job/JobList'));
+const JobDetailPage = lazy(() => import('../pages/Job/JobDetail'));
+import { jobDetailLoader, applyAction } from '../pages/Job/JobDetail';
 // Add JobDashboard
 const JobDashboardPage = lazy(() => import('../pages/dashboard/JobDashboard'));
